.......... [ZBXNEXT-8211] fixed compiler and iwyu warnings

+++ src/libs/zbxembed/global.c
@@ -69,8 +69,8 @@ static duk_ret_t	es_atob(duk_context *ctx)
 	if (SUCCEED != es_duktape_string_decode(duk_require_string(ctx, 0), &str))
 		return duk_error(ctx, DUK_RET_TYPE_ERROR, "cannot convert value to utf8");
 
-	buffer_size = (int)strlen(str) * 3 / 4 + 1;
-	buffer = zbx_malloc(buffer, (size_t)buffer_size);
+	buffer_size = strlen(str) * 3 / 4 + 1;
+	buffer = zbx_malloc(buffer, buffer_size);
 	str_base64_decode(str, buffer, buffer_size, &out_size);
 	duk_push_lstring(ctx, buffer, (duk_size_t)out_size);
 	zbx_free(str);
